feat: useFocusTimeout 훅으로 공통 모듈화
