+++ src/hooks/useFocusTimeout.tsx
@@ -1,0 +1,14 @@
+import { useEffect, useState } from 'react';
+
+export default function useFocusTimeout() {
+  const [isFocused, setIsFocused] = useState(false);
+  const handleFocus = () => setIsFocused(true);
+
+  useEffect(() => {
+    let focusTimeout: NodeJS.Timeout;
+    if (isFocused) focusTimeout = setTimeout(() => setIsFocused(false), 500);
+    return () => clearTimeout(focusTimeout);
+  }, [isFocused]);
+
+  return { isFocused, handleFocus };
+}
